pyproject.toml
[build-system]
requires = ["setuptools>=68", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "parallel-spectral-sensing-boards"
authors = [
    {name = "Flynn Marshall", email = "flyma@dtu.dk"}
]

description = "Python package for fast gathering of spectral sensor data using custom PCBs in parallel."
dynamic = ["version"]

readme = "README.md"
requires-python = ">=3.10"
license = {file="LICENSE"}
dependencies = [
    "requests>=2.26.0",
    "pandas>=2.0",
    "seaborn==0.12.2",
    "numpy>=1.23",
    "pyserial>=3.5",
    "matplotlib>=3.7",
    "pyyaml>=6.0",
    # NOTE: the 'north' package for N9 robot hardware is NOT on PyPI.
    # Install it separately following North Robotics instructions.
    # Without it, set robot.simulate: true in config.yaml to run without hardware.
]

classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

[project.scripts]
spectral-run = "spectral_board_manager.cli:main"
spectral-plotter = "spectral_board_manager.plotter:main"
experiment-run = "n9_controller.experiment_runner:main"

[project.optional-dependencies]
dev = [
  "ipython",
  "build",
  "twine",
]
lint = [
  "mypy",
  "ruff",
]

[project.urls]
repository = "https://github.com/team-capex/n9-spectral-sensing.git"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
n9_spectral_sensing_expansion = ["py.typed", "*.yaml"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
disallow_untyped_defs = true

[tool.setuptools_scm]
version_scheme = "post-release"
local_scheme = "no-local-version"